gcc 12 checks some things that clang doesn't, resulting in compile errors.
This series fixes some of theses issues, but still builds (and tests) with clang.

Unfortunately, we still don't have a clean gcc build due to an outstanding bug [1].

[1] https://gcc.gnu.org/bugzilla/show_bug.cgi?id=98056

Closes #10386

* github.com:scylladb/scylla:
  build: disable warnings that cause false-positive errors with gcc 12
  utils: result_loop: remove invalid and incorrect constraint
  service: forward_service: avoid using deprecated std::bind1st and std::not1
  repair: explicityl ignore tombstone gc update response
  treewide: abort() after switch in formatters
  db: view: explicitly ignore unused result
  compaction: leveled_compaction_strategy: avoid compares between signed and unsigned
  compaction_manager: compaction_reenabler: disambiguate compaction_state
  api: avoid function specialization in req_param
  alternator: ttl: avoid specializing class templates in non-namespace scope
  alternator: executor: fix signed/unsigned comparison in is_big()

+++ alternator/executor.cc
@@ -3087,7 +3087,7 @@ static void check_big_array(const rjson::value& val, int& size_left);
 
 static bool is_big(const rjson::value& val, int big_size = 100'000) {
     if (val.IsString()) {
-        return val.GetStringLength() > big_size;
+        return ssize_t(val.GetStringLength()) > big_size;
     } else if (val.IsObject()) {
         check_big_object(val, big_size);
         return big_size < 0;
